rstan and CRAN compatibility.
Since rstan is not yet in CRAN (and it is not expected there according to stan'
developer Ben Goodrich), using rstan as 'suggested' package is not possible,
neither calling its extract method (rstan::extract) to get the sampled values.
But using as.array works, so until rstan is not in CRAN, use 'as.array'.